+++ tests/Unit/BellGatewayTest.php
@@ -50,6 +50,26 @@ class BellGatewayTest extends Unit
         $this->tester->seeInDatabase('fs_foodsaver_has_bell', ['foodsaver_id' => $user2['id'], 'bell_id' => $bellId, 'seen' => 0]);
     }
 
+    public function testAddBellWithSameUserMultiplyTimes()
+    {
+        $user1 = $this->tester->createFoodsaver();
+        $user2 = $this->tester->createFoodsaver();
+        $bellData = Bell::create(
+            'bell should only exist once for each user',
+            $this->faker->text(50),
+            '',
+            [''],
+            [],
+            '',
+            true,
+        );
+
+        $this->gateway->addBellForUsers([$user1['id'], $user1['id'], $user2['id']], $bellData);
+        $bellId = $this->tester->grabFromDatabase('fs_bell', 'id', ['name' => $bellData->title, 'body' => $bellData->body]);
+        $this->tester->seeInDatabase('fs_foodsaver_has_bell', ['foodsaver_id' => $user1['id'], 'bell_id' => $bellId, 'seen' => 0]);
+        $this->tester->seeInDatabase('fs_foodsaver_has_bell', ['foodsaver_id' => $user2['id'], 'bell_id' => $bellId, 'seen' => 0]);
+    }
+
     public function testRemoveBellWorksIfIdentifierIsCorrect(): void
     {
         $this->tester->clearTable('fs_bell');