+++ CHANGELOG.md
@@ -84,6 +84,7 @@
 - Fix multiple network retries, possibly causing multiple identical sent chat and forum posts !4181 @DL6ER
 - Fix incorrect common working group links on user profiles !4171 @DL6ER
 - Fix no "load more messages" button in mailboxes on mobile view (`xs` screens) !4719 @DL6ER
+- Fix database problems with "fs_foodsaver_has_bell" when same user watchs forum thread multiply times !4267 @KoeMai
 - Fix closing of food basket requests #2361 !4178 @DL6ER
 - Fix error when user does not have a mailbox #2390 @KoeMai